Squirrel scripted position

+++ Sources/Athena2D/System.h
@@ -17,10 +17,10 @@ public:
 	virtual void init() = 0;
 	virtual void shutdown() = 0;
 
-	virtual void preload(std::shared_ptr<Entity> entity, double deltaTime) = 0;
-	virtual void beforeUpdate(double deltaTime) = 0;
-	virtual void update(std::shared_ptr<Entity> entity, double deltaTime) = 0;
-	virtual void afterUpdate(double deltaTime) = 0;
+	virtual void preload(std::shared_ptr<Entity> entity, float deltaTime) = 0;
+	virtual void beforeUpdate(float deltaTime) = 0;
+	virtual void update(std::shared_ptr<Entity> entity, float deltaTime) = 0;
+	virtual void afterUpdate(float deltaTime) = 0;
 
 	SystemState state = SystemState::INACTIVE;
